added changes to import transactions and save to db

+++ transactions/src/main/java/com/devstudio/transactions/composables/transactionList/TransactionItem.kt
@@ -189,7 +189,7 @@ private fun TransactionDate(
     appColors: AppColor,
 ) {
     Text(
-        text = DateFormatter.convertLongToDate(transaction.transactionDate.toLong()),
+        text = DateFormatter.convertLongToDate(transaction.transactionDate),
         color = appColors.material.onPrimaryContainer,
         fontSize = SECONDARY_TEXT_SIZE,
         modifier = Modifier.padding(end = 5.dp),
